+++ submitr/scripts/update_portal_object.py
@@ -15,10 +15,24 @@ import json
 import os
 import sys
 from typing import Callable, List, Optional, Tuple, Union
-from dcicutils.misc_utils import get_error_message, PRINT
-from dcicutils.portal_utils import Portal
 from dcicutils.command_utils import yes_or_no
 from dcicutils.common import ORCHESTRATED_APPS, APP_SMAHT
+from dcicutils.ff_utils import delete_metadata, purge_metadata
+from dcicutils.misc_utils import create_uuid, get_error_message, PRINT
+from dcicutils.portal_utils import Portal as PortalFromUtils
+
+
+class Portal(PortalFromUtils):
+
+    def delete_metadata(self, object_id: str) -> Optional[dict]:
+        if isinstance(object_id, str) and object_id and self.key:
+            return delete_metadata(obj_id=object_id, key=self.key)
+        return None
+
+    def purge_metadata(self, object_id: str) -> Optional[dict]:
+        if isinstance(object_id, str) and object_id and self.key:
+            return purge_metadata(obj_id=object_id, key=self.key)
+        return None
 
 
 _DEFAULT_APP = "smaht"
@@ -111,6 +125,7 @@ def main():
     parser.add_argument("--verbose", action="store_true", required=False, default=False, help="Verbose output.")
     parser.add_argument("--quiet", action="store_true", required=False, default=False, help="Quiet output.")
     parser.add_argument("--debug", action="store_true", required=False, default=False, help="Debugging output.")
+    parser.add_argument("--testify", action="store_true", required=False, default=False, help="Test-ify data.")
     args = parser.parse_args()
 
     def usage(message: Optional[str] = None) -> None:
@@ -139,19 +154,22 @@ def main():
         _post_or_patch_or_upsert(portal=portal,
                                  file_or_directory=args.post,
                                  explicit_schema_name=explicit_schema_name,
-                                 update_function=_post_from_file,
+                                 update_function=post_data,
+                                 update_action_name="POST",
                                  confirm=args.confirm, verbose=args.verbose, quiet=args.quiet, debug=args.debug)
     if args.patch:
         _post_or_patch_or_upsert(portal=portal,
                                  file_or_directory=args.patch,
                                  explicit_schema_name=explicit_schema_name,
-                                 update_function=_patch_from_file,
+                                 update_function=patch_data,
+                                 update_action_name="PATCH",
                                  confirm=args.confirm, verbose=args.verbose, quiet=args.quiet, debug=args.debug)
     if args.upsert:
         _post_or_patch_or_upsert(portal=portal,
                                  file_or_directory=args.upsert,
                                  explicit_schema_name=explicit_schema_name,
-                                 update_function=_upsert_from_file,
+                                 update_function=upsert_data,
+                                 update_action_name="UPSERT",
                                  confirm=args.confirm, verbose=args.verbose, quiet=args.quiet, debug=args.debug)
 
     if args.delete:
@@ -171,9 +189,56 @@ def main():
 
 
 def _post_or_patch_or_upsert(portal: Portal, file_or_directory: str,
-                             explicit_schema_name: str, update_function: Callable,
+                             explicit_schema_name: str,
+                             update_function: Callable, update_action_name: str,
                              confirm: bool = False, verbose: bool = False,
                              quiet: bool = False, debug: bool = False) -> None:
+
+    def is_schema_name_list(portal: Portal, keys: list) -> bool:
+        if isinstance(keys, list):
+            for key in keys:
+                if portal.get_schema(key) is None:
+                    return False
+                return True
+        return False
+
+    def post_or_patch_or_upsert(portal: Portal, file: str, schema_name: Optional[str],
+                                confirm: bool = False, verbose: bool = False,
+                                quiet: bool = False, debug: bool = False) -> None:
+
+        nonlocal update_function, update_action_name
+        if not quiet:
+            _print(f"Processing {update_action_name} file: {file}")
+        if data := _read_json_from_file(file):
+            if isinstance(data, dict):
+                if isinstance(schema_name, str) and schema_name:
+                    if debug:
+                        _print(f"DEBUG: File ({file}) contains an object of type: {schema_name}")
+                    update_function(portal, data, schema_name, confirm=confirm,
+                                    file=file, verbose=verbose, debug=debug)
+                elif is_schema_name_list(portal, list(data.keys())):
+                    if debug:
+                        _print(f"DEBUG: File ({file}) contains a dictionary of schema names.")
+                    for schema_name in data:
+                        if isinstance(schema_data := data[schema_name], list):
+                            if debug:
+                                _print(f"DEBUG: Processing {update_action_name}s for type: {schema_name}")
+                            for index, item in enumerate(schema_data):
+                                update_function(portal, item, schema_name, confirm=confirm,
+                                                file=file, index=index, verbose=verbose, debug=debug)
+                        else:
+                            _print(f"WARNING: File ({file}) contains schema item which is not a list: {schema_name}")
+                else:
+                    _print(f"WARNING: File ({file}) contains unknown item type.")
+            elif isinstance(data, list):
+                if debug:
+                    _print(f"DEBUG: File ({file}) contains a list of objects of type: {schema_name}")
+                for index, item in enumerate(data):
+                    update_function(portal, item, schema_name, confirm=confirm,
+                                    file=file, index=index, verbose=verbose, debug=debug)
+            if debug:
+                _print(f"DEBUG: Processing {update_action_name} file done: {file}")
+
     if os.path.isdir(file_or_directory):
         if ((files := glob.glob(os.path.join(file_or_directory, "*.json"))) and
             (files_and_schemas := _file_names_to_ordered_file_and_schema_names(portal, files))):  # noqa
@@ -183,133 +248,96 @@ def _post_or_patch_or_upsert(portal: Portal, file_or_directory: str,
                 if not (schema_name := file_and_schema[1]) and not (schema_name := explicit_schema_name):
                     _print(f"ERROR: Schema cannot be inferred from file name and --schema not specified: {file}")
                     continue
-                update_function(portal, file_and_schema[0], schema_name=schema_name,
-                                confirm=confirm, quiet=quiet, verbose=verbose, debug=debug)
+                post_or_patch_or_upsert(portal, file_and_schema[0], schema_name=schema_name,
+                                        confirm=confirm, quiet=quiet, verbose=verbose, debug=debug)
     elif os.path.isfile(file := file_or_directory):
-        if (not (schema_name := _get_schema_name_from_schema_named_json_file_name(portal, file)) and
-            not (schema_name := explicit_schema_name)):  # noqa
-            _print(f"ERROR: Schema cannot be inferred from file name and --schema not specified: {file}")
-            return
-        update_function(portal, file, schema_name=schema_name,
-                        quiet=quiet, verbose=verbose, debug=debug)
+        if ((schema_name := _get_schema_name_from_schema_named_json_file_name(portal, file)) or
+            (schema_name := explicit_schema_name)):  # noqa
+            post_or_patch_or_upsert(portal, file, schema_name=schema_name,
+                                    confirm=confirm, quiet=quiet, verbose=verbose, debug=debug)
+        else:
+            post_or_patch_or_upsert(portal, file, schema_name=schema_name,
+                                    confirm=confirm, quiet=quiet, verbose=verbose, debug=debug)
+            # _print(f"ERROR: Schema cannot be inferred from file name and --schema not specified: {file}")
+            # return
     else:
         _print(f"ERROR: Cannot find file or directory: {file_or_directory}")
 
 
-def _post_from_file(portal: Portal, file: str, schema_name: str,
-                    confirm: bool = False, verbose: bool = False,
-                    quiet: bool = False, debug: bool = False) -> bool:
-
-    def post_data(data: dict, index: int = 0) -> None:
-        nonlocal portal, file, schema_name, verbose, debug
-        if not (identifying_path := portal.get_identifying_path(data, portal_type=schema_name)):
+def post_data(portal: Portal, data: dict, schema_name: str, confirm: bool = False,
+              file: Optional[str] = None, index: int = 0,
+              verbose: bool = False, debug: bool = False) -> None:
+    if not (identifying_path := portal.get_identifying_path(data, portal_type=schema_name)):
+        if isinstance(file, str) and isinstance(index, int):
             _print(f"ERROR: Item for POST has no identifying property: {file} (#{index + 1})")
-            return
-        if portal.get_metadata(identifying_path, raise_exception=False):
-            _print(f"ERROR: Item for POST already exists: {identifying_path}")
-            return
-        if (confirm is True) and not yes_or_no(f"POST data for: {identifying_path} ?"):
-            return
-        if verbose:
-            _print(f"POST {schema_name} item: {identifying_path}")
-        try:
-            portal.post_metadata(schema_name, data)
-            if debug:
-                _print(f"DEBUG: POST {schema_name} item done: {identifying_path}")
-        except Exception as e:
-            _print(f"ERROR: Cannot POST {schema_name} item: {identifying_path}")
-            _print(get_error_message(e))
-            return
-
-    if not quiet:
-        _print(f"Processing POST file: {file}")
-    if data := _read_json_from_file(file):
-        if isinstance(data, dict):
-            post_data(data)
-        elif isinstance(data, list):
-            for index, item in enumerate(data):
-                post_data(item, index)
+        else:
+            _print(f"ERROR: Item for POST has no identifying property.")
+        return
+    if portal.get_metadata(identifying_path, raise_exception=False):
+        _print(f"ERROR: Item for POST already exists: {identifying_path}")
+        return
+    if (confirm is True) and not yes_or_no(f"POST data for: {identifying_path} ?"):
+        return
+    if verbose:
+        _print(f"POST {schema_name} item: {identifying_path}")
+    try:
+        portal.post_metadata(schema_name, data)
         if debug:
-            _print(f"DEBUG: Processing POST file done: {file}")
+            _print(f"DEBUG: POST {schema_name} item done: {identifying_path}")
+    except Exception as e:
+        _print(f"ERROR: Cannot POST {schema_name} item: {identifying_path}")
+        _print(get_error_message(e))
+        return
 
 
-def _patch_from_file(portal: Portal, file: str, schema_name: str,
-                     confirm: bool = False, verbose: bool = False,
-                     quiet: bool = False, debug: bool = False) -> bool:
-
-    def patch_data(data: dict, index: int = 0) -> None:
-        nonlocal portal, file, schema_name, verbose, debug
-        if not (identifying_path := portal.get_identifying_path(data, portal_type=schema_name)):
+def patch_data(portal: Portal, data: dict, schema_name: str, confirm: bool = False,
+               file: Optional[str] = None, index: int = 0,
+               verbose: bool = False, debug: bool = False) -> None:
+    if not (identifying_path := portal.get_identifying_path(data, portal_type=schema_name)):
+        if isinstance(file, str) and isinstance(index, int):
             _print(f"ERROR: Item for PATCH has no identifying property: {file} (#{index + 1})")
-            return
-        if not portal.get_metadata(identifying_path, raise_exception=False):
-            _print(f"ERROR: Item for PATCH does not already exist: {identifying_path}")
-            return
-        if (confirm is True) and not yes_or_no(f"PATCH data for: {identifying_path}"):
-            return
-        if verbose:
-            _print(f"PATCH {schema_name} item: {identifying_path}")
-        try:
-            portal.patch_metadata(identifying_path, data)
-            if debug:
-                _print(f"DEBUG: PATCH {schema_name} item OK: {identifying_path}")
-        except Exception as e:
-            _print(f"ERROR: Cannot PATCH {schema_name} item: {identifying_path}")
-            _print(e)
-            return
-
-    if not quiet:
-        _print(f"Processing PATCH file: {file}")
-    if data := _read_json_from_file(file):
-        if isinstance(data, dict):
-            patch_data(data)
-        elif isinstance(data, list):
-            for index, item in enumerate(data):
-                patch_data(item, index)
+        else:
+            _print(f"ERROR: Item for PATCH has no identifying property.")
+        return
+    if not portal.get_metadata(identifying_path, raise_exception=False):
+        _print(f"ERROR: Item for PATCH does not already exist: {identifying_path}")
+        return
+    if (confirm is True) and not yes_or_no(f"PATCH data for: {identifying_path}"):
+        return
+    if verbose:
+        _print(f"PATCH {schema_name} item: {identifying_path}")
+    try:
+        portal.patch_metadata(identifying_path, data)
         if debug:
-            _print(f"DEBUG: Processing PATCH file done: {file}")
+            _print(f"DEBUG: PATCH {schema_name} item OK: {identifying_path}")
+    except Exception as e:
+        _print(f"ERROR: Cannot PATCH {schema_name} item: {identifying_path}")
+        _print(e)
+        return
 
 
-def _upsert_from_file(portal: Portal, file: str, schema_name: Optional[str] = None,
-                      confirm: bool = False, verbose: bool = False,
-                      quiet: bool = False, debug: bool = False) -> bool:
-
-    def upsert_data(data: dict, index: int = 0) -> None:
-        nonlocal portal, file, schema_name, verbose, debug
-        if not (identifying_path := portal.get_identifying_path(data, portal_type=schema_name)):
+def upsert_data(portal: Portal, data: dict, schema_name: str, confirm: bool = False,
+                file: Optional[str] = None, index: int = 0,
+                verbose: bool = False, debug: bool = False) -> None:
+    if not (identifying_path := portal.get_identifying_path(data, portal_type=schema_name)):
+        if isinstance(file, str) and isinstance(index, int):
             _print(f"ERROR: Item for UPSERT has no identifying property: {file} (#{index + 1})")
-            return
-        existing_item = portal.get_metadata(identifying_path, raise_exception=False)
-        try:
-            if not existing_item:
-                if (confirm is True) and not yes_or_no(f"POST data for: {identifying_path} ?"):
-                    return
-                if verbose:
-                    _print(f"POST {schema_name} item: {identifying_path}")
-                portal.post_metadata(schema_name, data)
-            else:
-                if (confirm is True) and not yes_or_no(f"PATCH data for: {identifying_path} ?"):
-                    return
-                if verbose:
-                    _print(f"PATCH {schema_name} item: {identifying_path}")
-                portal.patch_metadata(identifying_path, data)
-            if debug:
-                _print(f"DEBUG: UPSERT {schema_name} item OK: {identifying_path}")
-        except Exception as e:
-            _print(f"ERROR: Cannot UPSERT {schema_name} item: {identifying_path}")
-            _print(e)
-            return
-
-    if not quiet:
-        _print(f"Processing UPSERT file: {file}")
-    if data := _read_json_from_file(file):
-        if isinstance(data, dict):
-            upsert_data(data)
-        elif isinstance(data, list):
-            for index, item in enumerate(data):
-                upsert_data(item, index)
+        else:
+            _print(f"ERROR: Item for UPSERT has no identifying property.")
+        return
+    exists = portal.get_metadata(identifying_path, raise_exception=False)
+    if ((confirm is True) and not yes_or_no(f"{'PATCH' if exists else 'POST'} data for: {identifying_path} ?")):
+        return
+    if verbose:
+        _print(f"{'PATCH' if exists else 'POST'} {schema_name} item: {identifying_path}")
+    try:
+        portal.post_metadata(schema_name, data) if not exists else portal.patch_metadata(identifying_path, data)
         if debug:
-            _print(f"DEBUG: Processing UPSERT file done: {file}")
+            _print(f"DEBUG: UPSERT {schema_name} item OK: {identifying_path}")
+    except Exception as e:
+        _print(f"ERROR: Cannot UPSERT {schema_name} item: {identifying_path}")
+        _print(e)
+        return
 
 
 def _create_portal(env: Optional[str] = None, app: Optional[str] = None,
@@ -392,6 +420,35 @@ def _get_schema(portal: Portal, name: str) -> Tuple[Optional[dict], Optional[str
                 if schema_name.replace("_", "").replace("-", "").strip().lower() == name.lower():
                     return schemas[schema_name], schema_name
     return None, None
+
+
+# Function (test-ify) to give all of the identifying properties
+# of the given Portal object new, test, unique values.
+def _testify(portal: Portal, data: dict, data_type: str) -> Optional[dict]:
+    _unique_test_uuid_map = {}
+    unique_test_uuid = create_uuid()
+    unique_test_identifier = f"test{unique_test_uuid.replace('-', '')}"
+    if not ((schema := portal.get_schema(data_type)) and (schema_properties := schema.get("properties"))):
+        return None
+    identifying_property_names = portal.get_identifying_property_names(schema)
+    for identifying_property_name in identifying_property_names:
+        if identifying_property_value := data.get(identifying_property_name):
+            if identifying_property_name == "uuid":
+                _unique_test_uuid_map[identifying_property_value] = unique_test_uuid
+                identifying_property_value = unique_test_uuid
+            elif isinstance(identifying_property_value, list):
+                for index, identifying_property_item in enumerate(identifying_property_value):
+                    identifying_property_value[index] = f"{identifying_property_item}{unique_test_identifier}"
+            else:
+                identifying_property_value = f"{identifying_property_value}{unique_test_identifier}"
+            data[identifying_property_name] = identifying_property_value
+    for schema_property_name in schema_properties:
+        if ((schema_property_name not in identifying_property_names) and
+            (schema_properties[schema_property_name].get("uniqueKey") is True)):  # noqa
+            if unique_property_value := data.get(schema_property_name):
+                unique_property_value = f"{unique_property_value}{unique_test_identifier}"
+                data[schema_property_name] = unique_property_value
+    return data
 
 
 def _print(*args, **kwargs) -> None: