+++ CHANGELOG.rst
@@ -6,6 +6,15 @@ smaht-submitr
 Change Log
 ----------
 
+
+
+1.0.0
+=====
+* Using pyinstaller to create a single independent executable (per command)
+  so commands can be run without having Python/pyenv/etc installed.
+* Some changes related to starting work on integration tests with the portal.
+
+
 0.8.3
 =====